flask_app/requirements.txt
dagshub==0.4.2
Flask==3.1.0
mlflow==2.19.0
mlflow_skinny==2.19.0
nltk==3.9.1
numpy==2.2.1
pandas==2.2.3
prometheus_client
python-dotenv==1.0.0
